+++ x/gov/migrations/v5/store.go
@@ -3,10 +3,10 @@ package v5
 import (
 	"cosmossdk.io/collections"
 	corestoretypes "cosmossdk.io/core/store"
+	govv1 "cosmossdk.io/x/gov/types/v1"
 
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	govv1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1"
 )
 
 var (